requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
cchardet==2.1.7
charset-normalizer==2.1.0
frozenlist==1.3.0
fuzzywuzzy==0.18.0
idna==3.3
lxml==4.9.1
multidict==6.0.2
numpy==1.23.1
pandas==1.4.3
psycopg2==2.9.3
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2022.1
six==1.16.0
snakeviz==2.1.1
soupsieve==2.3.2.post1
tornado==6.2
uvloop==0.16.0
yarl==1.7.2
